fix: Create separate billing details operations for transactions and subscriptions
