+++ UPGRADING.md
@@ -53,8 +53,9 @@ Usage of `paddle_billing.Resources.Reports.Operations.CreateReport` should be re
 ### 5. `BillingDetails` entity is no longer used for `billing_details` in request operations
 
 Usage of `paddle_billing.Entities.Shared.BillingDetails` for `billing_details` in request operations, should be replaced with:
-- `paddle_billing.Resources.Shared.Operations.Update.UpdateBillingDetails` for `UpdateTransaction` | `UpdateSubscription` | `PreviewUpdateSubscription`
-- `paddle_billing.Resources.Shared.Operations.Create.CreateBillingDetails` for `CreateTransaction`
+- `paddle_billing.Resources.Transactions.Operations.Create.CreateBillingDetails` for `CreateTransaction`
+- `paddle_billing.Resources.Transactions.Operations.Update.UpdateBillingDetails` for `UpdateTransaction`
+- `paddle_billing.Resources.Subscriptions.Operations.Update.UpdateBillingDetails` for `UpdateSubscription` | `PreviewUpdateSubscription`
 
 ## v0.2.0
