--- paddle_billing/Resources/Shared/Operations/Create/__init__.py
@@ -1,1 +1,0 @@
-from paddle_billing.Resources.Shared.Operations.Create.CreateBillingDetails import CreateBillingDetails
