--- paddle_billing/Resources/Shared/Operations/Update/__init__.py
@@ -1,1 +1,0 @@
-from paddle_billing.Resources.Shared.Operations.Update.UpdateBillingDetails import UpdateBillingDetails
