+++ paddle_billing/Resources/Subscriptions/Operations/Update/__init__.py
@@ -1,0 +1,2 @@
+from paddle_billing.Resources.Subscriptions.Operations.Update.SubscriptionDiscount import SubscriptionDiscount
+from paddle_billing.Resources.Subscriptions.Operations.Update.UpdateBillingDetails import UpdateBillingDetails
