+++ paddle_billing/Resources/Transactions/Operations/Update/__init__.py
@@ -1,0 +1,1 @@
+from paddle_billing.Resources.Transactions.Operations.Update.UpdateBillingDetails import UpdateBillingDetails
